NXT-57 Move things around
